# countries_openAPI.yaml
openapi: 3.0.0
info:
  title: Countries OpenAPI
  version: 1.0.0
servers:
- url: http://EC2AMAZ-0LEPQND.ngs.local:8080/
  description: Default Server.
paths:
...
  /regionalblock:
    post:
      operationId: addRegionalBlock
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/regionalblock'
      responses:
        default:
          content:
            '*/*':
              schema:
                type: object
          description: Default Response.
  /regionalblock/{id}:
    get:
      operationId: getRegionalBlock
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        schema:
          type: integer
      responses:
       '200':
        description: regionalblock returned successfully
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/regionalblock'
       '404':
          description: no content returned based on id provided
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
       '500':
        description: error contacting the database
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiResponse'
    put:
      operationId: updateRegionalBlock
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/regionalblock'
      responses:
          '204':
            description: regionalblock updated successfully
            content: 
              application/json:
                schema:
                  $ref: '#/components/schemas/ApiResponse'
          '400':
            description: error with data sent, response message may provide more detail
            content: 
              application/json:
                schema:
                  $ref: '#/components/schemas/ApiResponse'
          '500':
            description: error contacting the databsae
            content: 
              application/json:
                schema:
                  $ref: '#/components/schemas/ApiResponse'
    delete:
      operationId: deleteRegionalBlock
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        schema:
          type: integer
      responses:
        '204':
          description: regionalblock deleted successfully including any relationships
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: error with data sent, response message may provide more detail
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: error contacting the databsae
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  schemas:
...
    ApiResponse:
        title: ApiResponse
        type: object
        properties:
          code:
            type: integer
          type:
            type: string
          message:
            type: string
...
    regionalblock:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        acronym:
          type: string